Implementação do GetbyId e Validação de email

+++ server.js
@@ -5,6 +5,7 @@ import { databaseMemory } from "./database/database-memory.js";
 
 const server = fastify();
 const database = new databaseMemory();
+
 import 'dotenv/config'
 
 server.listen({
@@ -13,6 +14,11 @@ server.listen({
 
 console.log("Server is running");
 
+const validateEmail = (email) => {
+  const re = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+  return re.test(email);
+};
+
 // corrigir nome do parâmetro - feito
 server.post("/clients", (request, reply) => {
   const { name, email } = request.body;
@@ -20,8 +26,20 @@ server.post("/clients", (request, reply) => {
   // adicionar condições para verificar se body não está vazio, if's comuns já bastam - Feito
     if (name ===  "" ) return reply.status(400).send({status_code: 400, mensage: "Não é possível criar um nome vazio"})
 
-    
-  // verificar se o e-mail já existe
+  if (!email) {
+    return reply.status(400).send({ status_code: 400, message: "E-mail é obrigatório" });
+  }
+
+  if (!validateEmail(email)) {
+    return reply.status(400).send({ status_code: 400, message: "Formato de e-mail inválido" });
+  }
+
+
+  const emailExists = database.findByEmail(email);
+  if (emailExists) {
+    return reply.status(400).send({ status_code: 400, message: "E-mail já cadastrado" });
+  }
+
 
   database.create({
     name,
@@ -30,27 +48,29 @@ server.post("/clients", (request, reply) => {
 
   // retornar o cliente criado com o id
 
-  return reply.status(201).send({status_code: 201, mensage: "Cadastro Realizado!"});
+  return reply.status(201).send();
 });
 
 server.get("/clients", () => {
   const clients = database.list();
+  const client = database.getById(clientId);
 
-  return clients;
+  if (!client) {
+    return reply.status(404).send({ status_code: 404, message: "Cliente não encontrado" });
+  }
+
+  return client;
 });
 
-// método getById, para retornar apenas um cliente, caso não encontrar retornar uma mensagem de erro de não encontrado
-// server.getById("/clients/:id", () => {
-//   const clients = database.list();
 
-//   return clients;
-// });
-
-// verificar antes se o cliente existe e depois atualizar e verificar se o e-mail já existe para outro usuário (findByEmail)
-// tudo com as devidas mensagens de erro
 server.put("/clients/:id", (request, reply) => {
   const clientsID = request.params.id;
   const { name, email } = request.body;
+  const existingClient = database.getById(clientId);
+
+  if (!existingClient) {
+    return reply.status(404).send({ status_code: 404, message: "Cliente não encontrado" });
+  }
 
   database.update(clientsID, {
     name,
@@ -59,16 +79,20 @@ server.put("/clients/:id", (request, reply) => {
 
   // retornar o usuário atualizado com id
 
-  return reply.status(204).send({status_code: 204, mensage: "Cadastro Editado!"});
+  return reply.status(204).send();
 });
 
 // verificar se o usuário existe
 server.delete("/clients/:id", (request, reply) => {
   const clientsID = request.params.id;
 
+  const existingClient = database.getById(clientId);
+  if (!existingClient) {
+    return reply.status(404).send({ status_code: 404, message: "Cliente não encontrado" });
+  }
   database.delete(clientsID);
 
-  return reply.status(204).send({status_code: 204, mensage: "Cadastro Apagado!"});
+  return reply.status(204).send();
 });
 
 // como exportar requisições no postman
